we're encountering an issue with latest 4.3.3,
which seems to be caused by this change:
https://github.com/apache/karaf/pull/1427/commits/8cc2b247d0ae8b2c95439cf5a45118b8f0eb3b47
When connecting via SSH (Apache Mina client in our case), we expect a
return code for some of the commands we issue. The code still sends back
the command, but then invokes close with parameter immediate=true. This
leads to the return code almost never arriving at calling client.
If changed to session.close(false), the buffers get flushed and we
retrieve the return code again.